check level at engine discovery
